backend2 -  python/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
pymongo==4.5.0
gunicorn==21.2.0
requests==2.25.1
opentelemetry-api==1.6.0
opentelemetry-sdk==1.6.0
opentelemetry-instrumentation==0.25b0
opentelemetry-exporter-jaeger==1.6.0
opentelemetry-instrumentation-flask==0.25b0
opentelemetry-instrumentation-requests==0.25b0
opentelemetry-instrumentation-pymongo==0.25b0
